pyproject.toml
[project]
name = "api"
version = "1.0.0"
description = "The Garmin Activities api for Poseidon"
dependencies = [
    "blinker==1.9.0",
    "build==1.2.2.post1",
    "certifi==2024.8.30",
    "ciso8601==2.3.1",
    "click==8.1.7",
    "colorama==0.4.6",
    "Flask==3.1.0",
    "Flask-RESTful==0.3.10",
    "geojson==3.1.0",
    "grpcio==1.68.0",
    "influxdb3-python==0.9.0",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.4",
    "MarkupSafe==3.0.2",
    "numpy==2.1.3",
    "packaging==24.2",
    "pandas==2.2.3",
    "pyarrow==18.0.0",
    "pyproject_hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "reactivex==4.0.4",
    "setuptools==75.6.0",
    "six==1.16.0",
    "typing_extensions==4.12.2",
    "tzdata==2024.2",
    "urllib3==2.2.3",
    "Werkzeug==3.1.3"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"